fix: make exitableAt a uint32

+++ plasma_framework/contracts/mocks/exits/SpyPlasmaFrameworkForExitGame.sol
@@ -14,7 +14,7 @@ contract SpyPlasmaFrameworkForExitGame is PlasmaFramework {
     event EnqueueTriggered(
         uint256 vaultId,
         address token,
-        uint64 exitableAt,
+        uint32 exitableAt,
         uint256 txPos,
         uint256 exitId,
         address exitProcessor
@@ -31,7 +31,7 @@ contract SpyPlasmaFrameworkForExitGame is PlasmaFramework {
     function enqueue(
         uint256 _vaultId,
         address _token,
-        uint64 _exitableAt,
+        uint32 _exitableAt,
         PosLib.Position calldata _txPos,
         uint168 _exitId,
         IExitProcessor _exitProcessor
